[cmake,hdf5] move hdf5tests to test/hdf5
