+++ test/Runner.cpp
@@ -6,33 +6,36 @@
 // modification, are permitted under the terms of the BSD License. See
 // LICENSE file in the root of the Project.
 
-#include "TestH5.hpp"
-#include "TestEntityHDF5.hpp"
-#include "TestEntityWithMetadataHDF5.hpp"
-#include "TestEntityWithSourcesHDF5.hpp"
-#include "TestFileHDF5.hpp"
-#include "TestBlockHDF5.hpp"
-#include "TestSourceHDF5.hpp"
-#include "TestSectionHDF5.hpp"
-#include "TestDataSet.hpp"
-#include "TestH5Group.hpp"
-#include "TestDataArrayHDF5.hpp"
-#include "TestBaseTagHDF5.hpp"
-#include "TestMultiTagHDF5.hpp"
-#include "TestTagHDF5.hpp"
-#include "TestVariant.hpp"
-#include "TestValue.hpp"
-#include "TestPropertyHDF5.hpp"
 #include "TestNDArray.hpp"
 #include "TestUtil.hpp"
-#include "TestImplContainerHDF5.hpp"
-#include "TestDimensionHDF5.hpp"
-#include "TestFeatureHDF5.hpp"
-#include "TestDataAccessHDF5.hpp"
+#include "TestVariant.hpp"
+#include "TestValue.hpp"
 #include "TestOptionalObligatory.hpp"
+
 #include "TestValidate.hpp"
-#include "TestReadOnlyHDF5.hpp"
-#include "TestGroupHDF5.hpp"
+
+#include "hdf5/TestH5.hpp"
+#include "hdf5/TestEntityHDF5.hpp"
+#include "hdf5/TestEntityWithMetadataHDF5.hpp"
+#include "hdf5/TestEntityWithSourcesHDF5.hpp"
+#include "hdf5/TestFileHDF5.hpp"
+#include "hdf5/TestBlockHDF5.hpp"
+#include "hdf5/TestSourceHDF5.hpp"
+#include "hdf5/TestSectionHDF5.hpp"
+#include "hdf5/TestDataSet.hpp"
+#include "hdf5/TestH5Group.hpp"
+#include "hdf5/TestDataArrayHDF5.hpp"
+#include "hdf5/TestBaseTagHDF5.hpp"
+#include "hdf5/TestMultiTagHDF5.hpp"
+#include "hdf5/TestTagHDF5.hpp"
+#include "hdf5/TestPropertyHDF5.hpp"
+#include "hdf5/TestImplContainerHDF5.hpp"
+#include "hdf5/TestDimensionHDF5.hpp"
+#include "hdf5/TestFeatureHDF5.hpp"
+#include "hdf5/TestDataAccessHDF5.hpp"
+#include "hdf5/TestReadOnlyHDF5.hpp"
+#include "hdf5/TestGroupHDF5.hpp"
+
 #include "fs/TestAttributesFS.hpp"
 #include "fs/TestFileFS.hpp"
 #include "fs/TestBlockFS.hpp"
